move Uuid impl to seprate mod
