+++ rust/vote-tx-v2/src/gen_tx/tx_body.rs
@@ -2,8 +2,8 @@
 
 use minicbor::{Decode, Decoder, Encode, Encoder};
 
-use super::{EncodedCbor, EventMap, Uuid, Vote};
-use crate::Cbor;
+use super::{EncodedCbor, EventMap, Vote};
+use crate::{uuid::Uuid, Cbor};
 
 /// `TxBody` array struct length
 const TX_BODY_LEN: u64 = 4;
