+++ rust/vote-tx-v2/src/public_tx/prop_id.rs
@@ -2,7 +2,7 @@
 
 use minicbor::{Decode, Encode};
 
-use crate::gen_tx::Uuid;
+use crate::uuid::Uuid;
 
 /// A public voting proposal id struct.
 pub struct PropId(Uuid);
